feat: add a shuffle function
The user is able to click on a "shuffle" button that simulates making a number of randomized but
valid moves on the board. The reason we need to do it this way is that if one instead completely
randomize the values across the board, the game will become unsolvable in 50% of cases.

re #7